fix: Simplify code
