+++ posthog/temporal/common/slo_interceptor.py
@@ -1,10 +1,8 @@
-import traceback
 from typing import Any
 
 from django.conf import settings
 
 from temporalio import workflow
-from temporalio.exceptions import ActivityError, ApplicationError
 from temporalio.worker import (
     ExecuteWorkflowInput,
     Interceptor,
@@ -14,41 +12,7 @@ from temporalio.worker import (
 
 from posthog.slo.events import emit_slo_completed, emit_slo_started
 from posthog.slo.types import SloCompletedProperties, SloConfig, SloOutcome, SloStartedProperties
-
-
-def _unwrap_temporal_cause(exc: BaseException) -> BaseException:
-    """Unwrap Temporal's ``ActivityError → ApplicationError`` chain to the original cause.
-
-    When an activity raises, the Temporal SDK surfaces it on the workflow side
-    as ``ActivityError(cause=ApplicationError(...))``. This helper peels off
-    that wrapper so callers can inspect the real failure.
-    """
-    if isinstance(exc, ActivityError) and isinstance(exc.cause, ApplicationError):
-        return exc.cause
-    return exc
-
-
-def resolve_exception_class(exc: BaseException) -> str:
-    """Return the original exception class name, unwrapping Temporal's wrappers.
-
-    Use this from a workflow's own exception handler when you need to inspect
-    the failure *before* re-raising — e.g., to reclassify user-query errors as
-    a non-breach outcome before the SLO interceptor records the completion.
-    """
-    cause = _unwrap_temporal_cause(exc)
-    return getattr(cause, "type", None) or type(cause).__name__
-
-
-def _extract_error_trace(exc: BaseException, cause: BaseException) -> str:
-    """Return the best-available stack trace for a failed workflow.
-
-    Prefers the activity-side traceback that the activity wrapper stashes in
-    ``ApplicationError.details[0]`` (captured at the site of the real failure),
-    and falls back to a truncated workflow-side traceback otherwise.
-    """
-    if isinstance(cause, ApplicationError) and cause.details and isinstance(cause.details[0], str):
-        return cause.details[0]
-    return "\n".join(traceback.format_exception(exc)[:5])
+from posthog.temporal.common.errors import resolve_error_trace, unwrap_temporal_cause
 
 
 class _SloWorkflowInterceptor(WorkflowInboundInterceptor):
@@ -90,14 +54,11 @@ class _SloWorkflowInterceptor(WorkflowInboundInterceptor):
             return result
         except BaseException as exc:
             outcome = slo.outcome if slo.outcome is not None else SloOutcome.FAILURE
-            # Temporal wraps activity errors as ActivityError → ApplicationError;
-            # unwrap to get the original exception type, message, and trace.
-            # Workflows can override any of these by pre-populating completion_properties
-            # before re-raising (setdefault leaves existing values untouched).
-            cause = _unwrap_temporal_cause(exc)
+            # setdefault lets workflows pre-populate completion_properties to override any of these.
+            cause: BaseException = unwrap_temporal_cause(exc) or exc
             slo.completion_properties.setdefault("error_type", getattr(cause, "type", None) or type(cause).__name__)
             slo.completion_properties.setdefault("error_message", str(cause))
-            slo.completion_properties.setdefault("error_trace", _extract_error_trace(exc, cause))
+            slo.completion_properties.setdefault("error_trace", resolve_error_trace(exc))
             raise
         finally:
             duration_ms = (workflow.time() - start_time) * 1000
