+++ posthog/temporal/exports/types.py
@@ -1,30 +1,7 @@
-import typing
 import dataclasses
 from typing import Optional
 
-from temporalio.exceptions import ActivityError, ApplicationError
-
-
-class ExportErrorDetails(typing.NamedTuple):
-    exception_class: str | None = None
-    error_trace: str | None = None
-
-
-def extract_error_details(exc: BaseException) -> ExportErrorDetails:
-    """Extract failure metadata from a Temporal activity exception chain.
-
-    asyncio.gather(return_exceptions=True) yields BaseException, but Temporal
-    wraps activity failures as ActivityError -> ApplicationError. We narrow
-    through that chain to reach the structured details.
-    """
-    if not isinstance(exc, ActivityError) or not isinstance(exc.cause, ApplicationError):
-        return ExportErrorDetails()
-
-    cause = exc.cause
-    return ExportErrorDetails(
-        exception_class=cause.type,
-        error_trace=cause.details[0] if len(cause.details) >= 1 and isinstance(cause.details[0], str) else None,
-    )
+from posthog.temporal.common.errors import unwrap_temporal_cause
 
 
 @dataclasses.dataclass
@@ -44,3 +21,12 @@ class ExportAssetResult:
     exported_asset_id: int
     success: bool
     error: Optional[ExportError] = None
+
+
+def extract_error_details(exc: BaseException) -> ExportError | None:
+    """Build an ``ExportError`` from a failed activity exception, or ``None`` if there's nothing to attach."""
+    cause = unwrap_temporal_cause(exc)
+    if cause is None or not cause.type:
+        return None
+    error_trace = cause.details[0] if cause.details and isinstance(cause.details[0], str) else ""
+    return ExportError(exception_class=cause.type, error_trace=error_trace)
