+++ posthog/temporal/exports/workflows.py
@@ -8,7 +8,7 @@ from posthog.event_usage import EventSource
 from posthog.slo.types import SloConfig, SloOutcome
 from posthog.tasks.exports.failure_handler import is_user_query_error_type
 from posthog.temporal.common.base import PostHogWorkflow
-from posthog.temporal.common.slo_interceptor import resolve_exception_class
+from posthog.temporal.common.errors import resolve_exception_class
 from posthog.temporal.exports.activities import export_asset_activity
 from posthog.temporal.exports.retry_policy import EXPORT_RETRY_POLICY
 from posthog.temporal.exports.types import ExportAssetActivityInputs
